Fixed: Convert ImageManagementServices.xml minilang to groovy (OFBIZ-11604)
For an unknown reason (I did not dig in), in uploadProductImages
result.productId = parameters.productId must be set at start and not end
because parameters.productId does not exist at end. I guess because of the call
to addMultipleuploadForProduct, but as I said I did not dig in.

Also updateStatusImageManagement did not return productId

Unrelated removes an useless new line in ImageManagementServices.java