Fixed: Display rather than edit the From Date field of a Product Config Item Content when updating the content.
(OFBZ-12246)

When editing date-time fields, they are truncated to the second. However
the default From Date field value of a Product Config Item Content is
stored in milliseconds. The From Date is a primary key of the config
item content and is used when finding the relevant record to update.

When the From Date field was editable on the Edit Product Config Item
Content content form, it would be truncated to the second and included
in the update request POST to the server. Since the From Date was
altered, the relevant record to be updated could no longer be found by
its primary key.

By making the From Date field display only, we prevent truncation of its
value meaning the record to be updated can be found.