Fixed: Show WorkEffort names in FindWorkEffort page (OFBIZ-12253)
In FindWorkEffort page, "Work Effort Name" can be specified as search
option, but the result table do not show it.

It would be nice if WorkEffort names can be displayed. Actually,
'workEffortName' has already been added to the description of hyperlink
in 'workEffortId' field, but 'workEffortName' is not presented in
'fieldList', so names are failed to be shown.

Thanks: Xin Wang