Implemented : Add sub-screen decorator DetailScreenDecorator and EmbeddedDecorator (OFBIZ-11813)
To homogenize and help each theme to implement correctly the screen engine concept,
we add two sub-decorators like existing 'FindScreenDecorator'.

DetailScreenDecorator can be used to display a sum up of an object and
its related data.

Embedded Decorator can be used to inject a specific screen in an
structured page area.

Those are composed of several sections that can be handled by the
theme for specific display.

Thanks Nicolas