Fixed: Xml Element Link type hidden-form fail to resolve the form (OFBIZ-12260)
When you generate a link with the type hidden-form and the link present an update area, on the click jquery failed to resolve the form.

 Example of a current link generate :

  <a href="javascript:ajaxSubmitFormUpdateAreas('ListProductMeterActionMenus_delete_LF_1__0', 'embedded13009956-a841-47d3-86b1-b1931e6b430e,https://localhost:8443/catalog/control/ListProductMeters,productId=GZ-1000')">javascript:ajaxSubmitFormUpdateAreas('ListProductMeterActionMenus_delete_LF_1__0', 'embedded13009956-a841-47d3-86b1-b1931e6b430e,https://localhost:8443/catalog/control/ListProductMeters,productId=GZ-1000')</a>

The problem come from OfbizUtil.js:748

    var $form = jQuery("#" + form)

To solve it, we change the form resolution in the function ajaxSubmitFormUpdateAreas with use the form name instead the form id.

Thanks to Xin Wang for this suggest to fix it.