Improved: Improve js function OfbizUtil.ajaxSubmitFormUpdateAreas (OFBIZ-12268)
At this time the function ajaxSubmitFormUpdateAreas present on themes/common-theme/webapp/common/js/util/OfbizUtil.js take a form, submit it thought ajax and wait a success to analyse a json result for update an area with the cvsAreaString given on parameter.

I realized different improvements :

 * refactoring some js code to centralize the analyze of json result for check if the call is a success or error
 * Analyze the content-type of the response if is a json continue as before, else take the return and update first area (useful for submit search or filtering form)
 * Manage the case who the ajax call failed and we inform the user
 * Add a trigger to force the modal close on submit success

Thanks to Leila Mekika, Marie-Aline Pantais and Gil Portenseigne for looking on that.