Improved: Globalize the refresh on FindScreenDecorator (OFBIZ-11812) (OFBIZ-12261)
If you use a ComomnDynamicSingle form for realize a search, the FindScreenDecorator would be refresh all the container searchOptions + searchResult.

To keep the coherence, we move the screenStack resolution on the beginning of the decorator, and leave the searchResult to resolve it itself by usage of EmbeddedDecorator.

In practice to keep all dynamism on a search element :
  * Main screen with global decorator include the find screen wrap by main-decorator
  * The find screen include the search form and include the screen list result wrap by FindScreenDecorator
  * The screen result include the list result form wrap by EmbeddedDecorator