Fixed: Macro renderLink default height and width not retrieved on menus (OFBIZ-12279)
Menus link defaut parameters for height and width are not considered when the
link is generated. When parameter with empty value is passed to the macro, macro
default values are not retrieved.

To reproduce
you can create a "layered-modal" type link in a menu and click on the link
(without height nor width):
this will generate the link with parameters "data-dialog-height" and
"data-dialog-width" empty

Patch provided
When applied, the MacroMenuRenderer.renderLink  will not send empty values to
the macro so that default value will be retrieved (what is actually done on
MacroFormRenderer.renderLookup)

Thanks: Leila