Fixed: Improve js function OfbizUtil.ajaxSubmitFormUpdateAreas, some errors don't appear (OFBIZ-12268)
After the refactoring realized on issue OFBIZ-12268, a typo has been introduce when we scanned the presence of error during an ajax call.
The typo raised that ignore the case of an error is present as list.

The same typo is also present for event message list.

Thanks to Xin Wang to alert on this problem through issue OFBIZ-12272