Fixed: bug on button from catalog to party associated (OFBIZ-12289)
Problem when trying to access Party linked to Catalog.
When you click on a user associated with a catalog, the framework sends you to
party manager, but it's the currently logged in user that shows up in the
party manager, not the party that you want to see.

To reproduce, just go in the Catalog app, then on Catalogs, add a user to a
catalog, and try to access it via the button.

It comes from the fact that the parameter sent to the party manager via the
link doesn't have the right syntax.

Thanks: Gaetan Chaboussie