Improved: Propage product features from a marketing package (OFBIZ-12077)
When you create a product as a marketing package, you attempt when you ship this product to have the same constraint as you ship each product that compose it separately.

Marketing Package |  Composed product |  Product Feature
  GIFT_PACKAGE    |                   |
     1            |   GIZMO_HEAVY     |  (WEIGHT) HEAVY_SIZE
     2            |   GIZMO_TINY      |

On this example if I ordered a GIFT_PACKAGE, during the order process, currently you never see that you ordered a product with a feature HEAVY_SIZE, feature that can have some consequence for select a shipping method.

So when we need to analyze the features linked to a product, if the product is a marketing package, we propage the search to all products that compose it