Fixed: Setup instructions don't work (OFBIZ-12228)
The creation of a new company failed with constraint database error due to a missing functional General Ledger data.

To correct this, we separate the demo GlAccount demo data from Accounting demo data to allow it alone when we create the new company.

A second error can appear with the presence of a partyAcctgPreference if the validation isn't realized immediately, so we just add a control of the presence before call service createPartyAcctgPreference.

Thanks to Szabolcs Szekelyi for raise this issue and Jacques Leroux for its look