Fixed: ShoppingCart object does not recognize two products with different configurations (OFBIZ-12303)
By adding two equal products with different configurations in eCommerce cart it results in qty aggregation instead of adding 2 separate cart lines.

When we added a new configuration in cart, we analyze all configuration options already present to know if this configuration case is alredy load and merge the quantity added.

Thanks: Alexander Tzvetanov for raise this issue