Improved: Improve velocity of PartyHelper.getPartyName() with the cache
The class PartyHelper is massively used to resolve the name of a party without know is type.

     <field name="organizationPartyId" title="${uiLabelMap.ProductOrganization}">
         <display description="${groovy: org.apache.ofbiz.party.party.PartyHelper.getPartyName(delegator, organizationPartyId, true);} [${organizationPartyId}]"/>
     </field>

A party name have few change over time, or the PartyHelper.getPartyName function call the database at each call.

On huge screen, we distinctly clearly the database latency.

So for a cold data, no reason to didn't use the cache.