Improved: Refactor methods signature to reduce the number of params they use (OFBIZ-12335)
We currently have 115 checkstyle errors, most are related to methods using a too
high number of params.

Obviously nobody have currently time to work on this issue.

This commit increases the max ParameterNumber to 26 to hide all current related
errors. This reduces checkstyle errors to 54. It also allows to easier focus on
other errors.
It still possible to works on OFBIZ-12335 by temporary reverting this commit or
replacing max ParameterNumber by the number wanted (was 10, is 7 by default)