Improved: Refactor too long methods (OFBIZ-12338)
We currently have 54 checkstyle errors, few are related to too long methods

Obviously nobody have currently time to work on this issue.

This commit increases the max MethodLength to 784 to hide all current related
errors. This reduces checkstyle errors to 48. It also allows to easier focus on
other errors.

It still possible to works on OFBIZ-12338 by temporary reverting this commit or
replacing max ParameterNumber by the number wanted (was 500, is 60 by default)