Improved: Have the ability to revoke (expire) roles of a party (OFBIZ-5980)
Currently it only possible to delete a role. That can lead to potential error
messages, when a particular role is associated with records where the roleTypeId
is used as a foreign key in the entity-model/database and the user tries to
delete that role from the list.

In order to avoid such, it is better to revoke the PartyRole record by setting
an expiration date (end date) on the record.

Thanks: Pierre Smits