OFBIZ-5996 Improved: AgreementRole handling
Improved:
AddAgreementRole screen and NewArgreementRole form
When prior to adding a new agreement role (same or other) an agreement role was expired, the value for the fromDate field for the new agreement role was taken from the expired agreement role.
With this improvement, the fromDate is determined at the moment of loading the AddAgreementRole screen and passed on to the NewAgreementRole form.

Removed:
Accidentally included EditContactMech.groovy script.