Fixed: Quick Refund Entire Order no longer works (OFBIZ-12392)
This only fixes a 1st issue found while debugging.

Adds the order package to be able to debug Groovy in Eclipse. We need to add
packages to all Groovy scripts to be able to easily debug Groovy in Eclipse.

Adds .first() in updateReturnStatusFromReceipt at OrderReturnServices.groovy[506]
That is the fix

A lot of non functional formatting