Improved: Rename invoiceApplications menu-item label to Payments (OFBIZ-12405)
Application(s) is generic (also used within contexts elsewhere) in OFBiz.
Users have a better perception of context when looking at a menu-item Payments
in the screen of an invoice, than seeing Applications.
The additional benefit for developers is that with the implemented change the
tech come closer to the language of the user

jleroux: after a discussion with Pierre in the Jira, being more in favour of
using "Payments applications" (for instance the related screen is named
EditPaymentApplications), I had a look at
https://www.linguee.fr/francais-anglais/search?query=%22Payments%20applications%22
which was not much conclusive.

In a 1st attempt, using Google translate, I tried to translate few languages
based on a French translation:
        <value xml:lang="ar">طلبات المدفوعات </value>
        <value xml:lang="de">Zahlungsaufzeichnungen</value>
        <value xml:lang="en">Payments application</value>
        <value xml:lang="es">Registros de pago</value>
        <value xml:lang="fr">Enregistrement des paiements</value>
        <value xml:lang="hi-IN">अनुप्रयोग</value>
        <value xml:lang="it">Registri di pagamento</value>

I finally think that Pierre's is right and we could simply use Payments there
rather than Applications and even "Payments applications" (could be debated...)

So I just changed the French and English labels and the labels names of
AccountingApplicationSuccess to AccountingPaymentApplicationSuccess
and AccountingApplications to AccountingPaymentsApplications

I let the users of other languages to decide what to put there or discuss if
they see an interest.

Thanks: Pierre for report and suggestion