Improved: VIEW permissions - invoice overview term triggers (OFBIZ-12429)
Currently, when using a userId with only VIEW permission, the section terms shows triggers to term line items in the invoice overview. This should not be, as those triggers on the line items are reserved for users with CREATE/UPDATE permissions.
See (test with): https://localhost:8443/accounting/control/invoiceOverview?invoiceId=demo10001

Modified:
invoiceScreens.xml.xml
- adjusted form ref to grid ref for invoice term records