Improved: VIEW permissions and Payment Applications (OFBIZ-12427)
Currently, a user with only 'VIEW' permissions, as demonstrated in trunk demo with userId = auditor, accessing the Payment Applications screen on a payment, sees editable fields and/or triggers (to requests) reserved for users with 'CREATE' or 'UPDATE' permissions.
This can be observed/tested via:
https://demo-trunk.ofbiz.apache.org/accounting/control/editPaymentApplications?paymentId=8004
https://demo-trunk.ofbiz.apache.org/accounting/control/editPaymentApplications?paymentId=8003
etc.

modified: PaymentScreens.xml
restructured screen EditPaymentApplications in order to work for users with VIEW and/or CREATE/UPDATE
permissions, additional cleanup