Improved: Have configuration options for Content (OFBIZ-6670)
The configuration of an OFBiz is determined by 2 aspects:
configuration settings required to successfully start the implementation
configuration settings required to have component functions work properly
The latter involve default choices for various functions to cater to the usability of business processes (and therefore screen/forms and/or screens/templates.
Several of these configuration settings are already loaded into the SystemProperty table in the underlying RDBMS, to enable adjustments of the systemPropertyValue while the implementation is in use. While it is not expected that these settings change often, it is possible that they are required to be changed during the lifespan of the company/organisation.
However, it should not be required to go through a deployment process of regarding such changed configuration settings in the OFBiz production instance (which most - if not all - companies/organisations require when files are changes).
Having all configuration settings of components in the SystemProperty table enhances the usability of OFBiz allowing system administrators to adjust the value(s) of system property records while the instance keeps running.

added: SystemProperty records in ContentSystemPropertyData.xml
Modified:
ofbiz-component.xml: reference of ContentSystemPropertyData.xml to load at seed loading moment