Fixed: Weird behaviour with currency symbols (OFBIZ-12443)
In some screens the $ currency symbol is not showing well. I guess it's true for all screens so I put all applications and plugins. I don't know if only trunk is affected.

Modified: FinAccountForms.xml
ensured that for the monetary amount fields (actualBalance and availableBalance) the correct currencyUomId value is taken