Improved: Have a Menu in Manufacturing featuring actions to create the main objects (OFBIZ-11782)
Currently the create buttons for the main objects of the manufacturing application are located within find and profile widgets/templates of those objects.
In order to improve the usability of this application, OFBiz and thus the appeal of it for adopters and users, these create buttons/links/etc. should be in a main action menu visible at all times when a user is working within the appliclation of the component.

modified:
ManufacturingMenus.xml - added MainActionMenu
CommonScreens.xlm - added pre-body decorator-section with MainActionMenu ref.