Improved: Move edit of invoice header to InvoiceSubTabBar (OFBIZ012440)
In general, a <object>TabBar menu is intended to give users access to overview related to the <object>. And a <object>SubTabBar menu is intended to give users to actions to be performed on the <object>, like status changes, adding new records, etc.
The menu-item for editing the invoice header (aptly translated as 'Header') is currently located in the InvoiceTabBar menu in AccountingMenus.xml.
Given that this menu-item is intended for users with UPDATE permissions and to give those users access to a screen to edit the data, this menu-item must move to the InvoiceSubTabBar menu

Modified: AccountingMenus.xml
- moved 'edit' menu-item for invoice from InvoiceTabBar menu to InvoiceSubTabBar menu. adjusted label.