Fixed: Price Rules: Price rules configurations are not syncing with eCommerce storefront (OFBIZ-12450)
When adding a Price Rule, whether it be a Product Category, Product, Party, Party Classification, the system doesn't show a price difference.
The price rules work with the Database as they do get changed when saved. But their functionality doesn't work, it won't give a price increase or decrease or restriction.

Modified:
Various data files to enhance demonstration of price rule functionality