Improved: Fix some bugs Spotbugs reports (OFBIZ-12386)
SpotBugs reports that
<<org.apache.ofbiz.base.util.collections.FlexibleServletAccessor.equals(Object)
does not check for parameter nullity

This implementation of equals(Object) does not respect the contract defined by
java.lang.Object.equals() because it does not check the nullity of the parameter
received as argument. All equals() methods must return false when they receive
a null value.>>

This fixes it, but despite SpotBugs continues to report the same error, excludes
it.

Also removes unused expandContext private var